change: updated stories file with better example
